Fix lua script example.
Former-commit-id: ec7061d9be097537f27e6067c7f13d91d758d256

+++ example/lua-scripts/src/Main.cpp
@@ -45,13 +45,13 @@ int main(int argc, char** argv)
         ->registerParser<file::LuaScriptParser>("lua")
         ->queue("script/main.lua");
 
-	Signal<Canvas::Ptr, uint, uint>::Slot nextFrame;
+	Signal<Canvas::Ptr, uint, float>::Slot nextFrame;
 	Signal<file::AssetLibrary::Ptr>::Slot loaded = sceneManager->assets()->complete()->connect(
 		[&](file::AssetLibrary::Ptr assets)
 		{
 			loaded = nullptr;
 
-			nextFrame = canvas->enterFrame()->connect([&](Canvas::Ptr, uint, uint)
+			nextFrame = canvas->enterFrame()->connect([&](Canvas::Ptr, uint, float)
 			{
 				nextFrame = nullptr;
 				root->addComponent(sceneManager->assets()->script("script/main.lua"));
@@ -61,7 +61,7 @@ int main(int argc, char** argv)
 
 	sceneManager->assets()->load();
 
-	auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint time, uint deltaTime)
+	auto enterFrame = canvas->enterFrame()->connect([&](Canvas::Ptr canvas, uint time, float deltaTime)
 	{
 		sceneManager->nextFrame();
 	});
